Updated scripts to pull database credentials from /etc/db.conf. Simplifies db operations.

+++ cgi-bin/clear.php
@@ -1,14 +1,19 @@
 <?php
-    # store this in  /etc/pass
-    $passfile = fopen("../../etc/pass", "r") or die("TODO:send http error");
-    
-    #$user = key(
-
-    $suppliedPass = key($_POST);
-    if (!strcmp($realPass, $suppliedPass)) {
-        $conn = new mysqli('simpledb.db', $user, $pass, 'simple_server');
-        if (!$conn->connect_error) {
-            mysqli_query($conn, "DELETE FROM users;");
-        } 
+    include 'util.php';
+    if (!($auth = authenticate())) {
+        header("HTTP/1.1 500 Internal Server Error");
+        return;
+    } 
+    # Authentication failure
+    if ($auth['passwd'] != $_POST['passwd']) {
+        http_response_code(400);
+        header("HTTP/1.1 400 Bad Request");
+        return;
+    }
+    # Attempt to connect, if successful delete user table.
+    $conn = sql_connect();
+    if ($conn != FALSE && !$conn->connect_error) {
+        mysqli_query($conn, "DELETE FROM users;");
+        header("HTTP/1.1 200 OK");
     }
 ?>
